services/ear/tcp_transcriber/requirements.txt
openai-whisper == 20250625
torchaudio ~= 2.7.0
